crypt: add region module
